Primeira versão do PetBook para empresa Bem Banhado Pet Shop.

+++ client_window.py
@@ -1,27 +1,37 @@
 import sys
 from PyQt5.QtWidgets import QApplication, QWidget, QVBoxLayout, QHBoxLayout, QTableWidget, QTableWidgetItem, \
-    QPushButton, QDialog, QFormLayout, QLineEdit, QCheckBox, QComboBox, QDateEdit
+    QPushButton, QDialog, QFormLayout, QLineEdit, QCheckBox, QComboBox, QLabel, QMessageBox, QPlainTextEdit
 from PyQt5.QtCore import Qt, QDate
-from lojaDB import buscar_compras, registrar_compra, editar_compra, excluir_compra
+from lojaDB import buscar_compras, registrar_compra, editar_compra, excluir_compra, buscar_cliente_por_id
 from utils import criar_seletor_data
 
 class ClientWindow(QWidget):
     def __init__(self, client_id):
         super().__init__()
         self.client_id = client_id
-        self.setWindowTitle('Detalhes do Cliente')
-        self.setGeometry(100, 100, 900, 600)
+        self.client_name = self.get_client_name()
+        self.setWindowTitle('Ficha do Cliente')
+        self.setGeometry(100, 100, 1000, 600)
         self.initUI()
+
+        #expand
+        #self.showMaximized()
 
     def initUI(self):
         main_layout = QVBoxLayout()
 
+        # Labels para ID e Nome do Cliente
+        self.label_id = QLabel(f"ID do Cliente: {self.client_id}")
+        self.label_nome = QLabel(f"Nome do Cliente: {self.client_name}")
+
+        main_layout.addWidget(self.label_id)
+        main_layout.addWidget(self.label_nome)
+
         # Lista de compras
         self.table_compras = QTableWidget()
-        self.table_compras.setColumnCount(10)  # Atualizar o número de colunas
+        self.table_compras.setColumnCount(6)  # Atualizar o número de colunas
         self.table_compras.setHorizontalHeaderLabels(
-            ["Número Item", "ID", "Data da Venda", "Produto", "Valor da Venda", "Modo de Pagamento", "Data de Pagamento",
-             "Data de Envio", "Código de Rastreio", "Enviado?"])
+            ["Número do Serviço", "ID", "Data do Agendamento", "Serviço", "Valor da Venda", "Notas"])
         self.table_compras.setSelectionBehavior(QTableWidget.SelectRows)
         self.table_compras.setEditTriggers(QTableWidget.NoEditTriggers)
 
@@ -49,30 +59,13 @@ class ClientWindow(QWidget):
         self.load_compras()
         self.table_compras.resizeColumnsToContents()  # Ajustar após carregar os dados
 
-        # Aplicar estilo diretamente
-        self.setStyleSheet("""
-                            QWidget {
-                                background-color: #f0f0f0;
-                            }
 
-                            QHeaderView::section {
-                                background-color: #f0f0f0; 
-                                color: black; 
-                            }
-                            QPushButton {
-                                background-color: #3c4c7d;
-                                color: white;
-                                border-radius: 5px;
-                                padding: 10px;
-                                font-size: 16px;
-                            }
-                            QPushButton:hover {
-                                background-color: #24346c;
-                            }
-                            QPushButton:pressed {
-                                background-color: #465184;
-                            }
-                        """)
+
+    def get_client_name(self):
+        client_data = buscar_cliente_por_id(self.client_id)
+        if client_data:
+            return client_data[1]  # Assume que o nome está na segunda posição
+        return "Desconhecido"
 
     def load_compras(self):
         try:
@@ -80,21 +73,28 @@ class ClientWindow(QWidget):
             print(f"Compras carregadas: {compras}")
             self.table_compras.setRowCount(len(compras))
             for row_idx, compra in enumerate(compras):
-                for col_idx, value in enumerate(compra[:-1]):
+                for col_idx, value in enumerate(compra[:5]):
                     item = QTableWidgetItem(str(value))
                     self.table_compras.setItem(row_idx, col_idx, item)
 
-                enviado_checkbox = QTableWidgetItem()
-                enviado_checkbox.setFlags(Qt.ItemIsUserCheckable | Qt.ItemIsEnabled)
-                enviado_checkbox.setCheckState(Qt.Checked if compra[-1] else Qt.Unchecked)
-                self.table_compras.setItem(row_idx, len(compra) - 1, enviado_checkbox)
+                    # Carregar o campo de "Notas" (último campo)
+                    notas_item = QTableWidgetItem(compra[5])  # Acessar o índice correto para as notas
+                    self.table_compras.setItem(row_idx, 5, notas_item)  # Colocar as notas na última coluna
+
+                #enviado_checkbox = QTableWidgetItem()
+                #enviado_checkbox.setFlags(Qt.ItemIsUserCheckable | Qt.ItemIsEnabled)
+                #enviado_checkbox.setCheckState(Qt.Checked if compra[-1] else Qt.Unchecked)
+                #self.table_compras.setItem(row_idx, len(compra) - 1, enviado_checkbox)
         except Exception as e:
             print(f"Erro ao carregar compras: {e}")
 
     def adicionar_compra(self):
         dialog = CompraDialog(self, self.client_id)
         if dialog.exec_():
+            print("salvando a nova compra...")
+            dialog.save_compra()
             self.load_compras()
+            print("Nova compra adicionada com sucesso.")
 
     def editar_compra(self):
         selected_row = self.table_compras.currentRow()
@@ -119,11 +119,10 @@ class ClientWindow(QWidget):
                                 dados_atualizados['data_venda'],
                                 dados_atualizados['produto'],
                                 dados_atualizados['valor_venda'],
-                                dados_atualizados['modo_pagamento'],
-                                dados_atualizados['data_pagamento'],
-                                dados_atualizados.get('data_envio', None),
-                                dados_atualizados.get('codigo_rastreio', None),
-                                1 if dados_atualizados.get('enviado', False) else 0
+                                #dados_atualizados['modo_pagamento'],
+                                #dados_atualizados['data_pagamento'],
+                                #dados_atualizados.get('data_envio', None),
+                                dados_atualizados.get('notas', None),
                             )
                             # Recarregar as compras para refletir as mudanças
                             self.load_compras()
@@ -138,12 +137,27 @@ class ClientWindow(QWidget):
         selected_row = self.table_compras.currentRow()
         if selected_row != -1:
             compra_id = self.table_compras.item(selected_row, 0).text()  # Ajustar para pegar o ID correto
-            try:
-                excluir_compra(compra_id)
-                print(f"Compra com ID {compra_id} excluída com sucesso.")
-                self.load_compras()
-            except Exception as e:
-                print(f"Erro ao excluir compra: {e}")
+            #Caixa de confirmação
+            msg_box = QMessageBox()
+            msg_box.setIcon(QMessageBox.Warning)
+            msg_box.setText("Tem certeza que deseja excluir esta compra?")
+            msg_box.setInformativeText(f"A compra com ID {compra_id} será permanentemente apagada")
+            msg_box.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
+            msg_box.setDefaultButton(QMessageBox.No)
+
+            #Exibir mensagem e capturar a resposta
+            response = msg_box.exec_()
+            if response == QMessageBox.Yes:
+                try:
+                    excluir_compra(compra_id)
+                    print(f"Compra com ID {compra_id} excluída com sucesso.")
+                    self.load_compras()
+                except Exception as e:
+                    print(f"Erro ao excluir compra: {e}")
+            else:
+                print(f"Exclusão da compra cancelada.")
+        else:
+            print("Nenhuma compra selecionada para exclusão")
 
 
 class CompraDialog(QDialog):
@@ -160,38 +174,46 @@ class CompraDialog(QDialog):
         self.input_data_venda = criar_seletor_data()
         self.input_produto = QLineEdit()
         self.input_valor_venda = QLineEdit()
-        self.input_modo_pagamento = QComboBox()
-        self.input_modo_pagamento.addItems(["Pix", "Débito", "Crédito", "Dinheiro"])
-        self.input_data_pagamento = criar_seletor_data()
-        self.input_data_envio = criar_seletor_data()
-        self.input_codigo_rastreio = QLineEdit()
-        self.checkbox_enviado = QCheckBox()
+        #self.input_modo_pagamento = QComboBox()
+        #self.input_modo_pagamento.addItems(["Pix", "Débito", "Crédito", "Dinheiro"])
+        #self.input_data_pagamento = criar_seletor_data()
+        #self.input_data_envio = criar_seletor_data()
+        self.input_notas = QPlainTextEdit()
+        #self.checkbox_enviado = QCheckBox()
 
         if self.compra_data:
             try:
                 self.input_data_venda.setDate(QDate.fromString(str(self.compra_data[2]), 'dd-MM-yyyy'))
                 self.input_produto.setText(self.compra_data[3])
                 self.input_valor_venda.setText(str(self.compra_data[4]))
-                self.input_modo_pagamento.setCurrentText(self.compra_data[5])
-                self.input_data_pagamento.setDate(QDate.fromString(str(self.compra_data[6]), 'dd-MM-yyyy'))
-                self.input_data_envio.setDate(QDate.fromString(str(self.compra_data[7]), 'dd-MM-yyyy'))
-                self.input_codigo_rastreio.setText(self.compra_data[8])
-                self.checkbox_enviado.setChecked(bool(self.compra_data[9]))  # Verifica o valor da coluna "Enviado?"
+                #self.input_modo_pagamento.setCurrentText(self.compra_data[5])
+                #self.input_data_pagamento.setDate(QDate.fromString(str(self.compra_data[6]), 'dd-MM-yyyy'))
+                #self.input_data_envio.setDate(QDate.fromString(str(self.compra_data[7]), 'dd-MM-yyyy'))
+                self.input_notas.setPlainText(self.compra_data[5])
+                #self.checkbox_enviado.setChecked(bool(self.compra_data[9]))  # Verifica o valor da coluna "Enviado?"
             except Exception as e:
                 print(f"Erro ao carregar dados da compra: {e}")
 
-        layout.addRow("Data da Venda:", self.input_data_venda)
-        layout.addRow("Produto:", self.input_produto)
+        layout.addRow("Data do Agendamento:", self.input_data_venda)
+        layout.addRow("Serviço:", self.input_produto)
         layout.addRow("Valor da Venda:", self.input_valor_venda)
-        layout.addRow("Modo de Pagamento:", self.input_modo_pagamento)
-        layout.addRow("Data de Pagamento:", self.input_data_pagamento)
-        layout.addRow("Data de Envio:", self.input_data_envio)
-        layout.addRow("Código de Rastreio:", self.input_codigo_rastreio)
-        layout.addRow("Enviado?", self.checkbox_enviado)
+        #layout.addRow("Modo de Pagamento:", self.input_modo_pagamento)
+        #layout.addRow("Data de Pagamento:", self.input_data_pagamento)
+        #layout.addRow("Data de Envio:", self.input_data_envio)
+        layout.addRow("Notas:", self.input_notas)
+        #layout.addRow("Enviado?", self.checkbox_enviado)
 
-        self.button_save = QPushButton("Salvar")
-        self.button_save.clicked.connect(self.save_compra)
-        layout.addWidget(self.button_save)
+        button_layout = QHBoxLayout()
+        self.button_ok = QPushButton("OK")
+        self.button_cancel = QPushButton("Cancelar")
+
+        self.button_ok.clicked.connect(self.accept)
+        self.button_cancel.clicked.connect(self.reject)
+
+        button_layout.addWidget(self.button_ok)
+        button_layout.addWidget(self.button_cancel)
+
+        layout.addRow(button_layout)
 
         self.setLayout(layout)
 
@@ -199,43 +221,43 @@ class CompraDialog(QDialog):
         return {
             'data_venda': self.input_data_venda.date().toString('dd-MM-yyyy'),
             'produto': self.input_produto.text(),
-            'valor_venda': float(self.input_valor_venda.text().replace(',', '.')) if self.input_valor_venda.text() else 0.0,
-            'modo_pagamento': self.input_modo_pagamento.currentText(),
-            'data_pagamento': self.input_data_pagamento.date().toString('dd-MM-yyyy'),
-            'data_envio': self.input_data_envio.date().toString('dd-MM-yyyy') if self.input_data_envio.date() != QDate.currentDate() else None,
-            'codigo_rastreio': self.input_codigo_rastreio.text(),
-            'enviado': self.checkbox_enviado.isChecked()
+            'valor_venda': self.input_valor_venda.text(),
+            #'modo_pagamento': self.input_modo_pagamento.currentText(),
+            #'data_pagamento': self.input_data_pagamento.date().toString('dd-MM-yyyy'),
+            #'data_envio': self.input_data_envio.date().toString('dd-MM-yyyy') if self.input_data_envio.date().isValid() else None,
+            'notas': self.input_notas.toPlainText(),
+            #'enviado': self.checkbox_enviado.isChecked()
         }
 
     def save_compra(self):
         try:
             dados_compra = self.get_dados_compra()
-            print(f"Salvando compra: {dados_compra}")
+            print(f"salvando compra: {dados_compra}")  # Log para ver os dados obtidos
             if self.compra_data:
                 editar_compra(
-                    self.compra_data[1],  # ID da compra
+                    self.compra_data[1],  # id da compra
                     dados_compra['data_venda'],
                     dados_compra['produto'],
                     dados_compra['valor_venda'],
-                    dados_compra['modo_pagamento'],
-                    dados_compra['data_pagamento'],
-                    dados_compra.get('data_envio', None),
-                    dados_compra.get('codigo_rastreio', None),
-                    1 if dados_compra.get('enviado', False) else 0
+                    #dados_compra['modo_pagamento'],
+                    #dados_compra['data_pagamento'],
+                    #dados_compra.get('data_envio', None),
+                    dados_compra.get('notas', None),
+                    #1 if dados_compra.get('enviado', False) else 0
                 )
+                print("Compra editada com sucesso.")  # Verifique se chega aqui ao editar
             else:
+                print("Tentando registrar nova compra...")  # Verifique se o fluxo chega aqui
                 registrar_compra(self.client_id, **dados_compra)
 
             self.accept()
-        except ValueError as ve:
-            print(f"Erro: valor da venda inválido '{dados_compra['valor_venda']}'. Detalhes: {ve}")
-            # Adicionar uma mensagem de erro na interface se desejado
         except Exception as e:
             print(f"Erro ao salvar compra: {e}")
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     app = QApplication(sys.argv)
-    client_window = ClientWindow(1)
+    client_id = 1  # Defina o ID do cliente
+    client_window = ClientWindow(client_id)
     client_window.show()
     sys.exit(app.exec_())
